backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
openai>=1.20.0
python-dotenv>=1.0.1
twilio>=8.13.0
fpdf2>=2.7.9
pydantic>=2.6.0
python-multipart>=0.0.9
httpx>=0.27.0
